frontend/ renaming participantList and customChat
